implemented 10a

+++ src/main.rs
@@ -87,7 +87,7 @@ fn solve_day9() -> (usize, usize) {
     )
 }
 
-fn solve_day10() -> (usize, usize) {
+fn solve_day10() -> (i32, usize) {
     let input = fs::read_to_string("day10.txt").expect("Should have been able to read day10.txt");
     (
         day10::sum_signal_strengths(&input, 20, 40),
